drawing the axes
The -update- function takes the -items- array from the API response. Our
bar chart will have a vertical bar for each repository, so we create the
horizontal -xScale- using the -scaleBand- helper to evenly space the bars

 - The -.domain- is the full_name from each repository.
 - Each repository's full name is unique, so this will result in 20 bands.

The vertical -yScale- is used to visualize the number of stars that each
repository has, so its domain goes form zero to the max -stargazers_count-

We use -nice- here to round the top of the scale to the next tick value.

After creating the scales, we create the axis generators and then use those
generations to draw the axes to the containers.

The last thing to do is call our -update- function from inside the -d3.json-
callback, passing the -items- array.

We
are able to go straight from fetching the data to calling -update- because the
GitHub Search API conveniently returns the data in the format we need for rendering.