Drawing the bars
Drawing a bunch of 'rect' elements. The key function of -.data-
extracts the full_name property, which is being used as the unique
identifier for each repository.

For now, using the simple -.join- method without separate handling
for entering, updating, and exiting elements (will come later)

The -x- attribute is set based on looking up the -full_name- in xScale,
and the width is based on the -bandwidth- method on -xScale.

The -y- and -height- attributes are a bit trickier this time and require
some explanation.

Looking at the definition of -yScale- , we see that the domain is:
- [0, d3.max(items, d => d.stargazers_count)] and the range is [height -
   margin.bottom, margin.top]
With the values we've set, that range expands to [380, 20]

The range goes from high to low, meaning that high values in the domain
map to lower values in the range, and vice versa.
- This is because -y- values in computer graphics count down from the top
  of the screen, but in our graph we want -y- values to count up from the
  bottom of the graph.

The other thing that makes this ticky is that SVG rectangles are drawn from
the top-left corner, which will likely be different for each bar, so we need
to set a variable height that makes all the bars hit the bottom axis.

Because of all this, we set the -y- attribute of the bar to -yScale(d.stargazers_count)-
which gives the vertical position of the top bar.
- To calcultae the height of the bar, we use -yScale(0) - yScale(d.stargazers_count)-

Cals
Calling -yScale(0)- gives the vertical position of the bottom of the bar (all
the bars should have their base at 0 in the domain), so substracting the position
of the top of the bar from the position of the bottom of the bar gives the height
of the bar.

- We need to end up with a positive height, so we have to subtract the smaller number
  from the larger number.
  - The top of the bar is a smaller number is the display range, even though it's a
    larger number in the domain.

- As you look at the bars, remember that each bar is drawn from its top-left corner,
  and that the heights are calculated such that the bottoms of all bars align.