Updating the sidebar on hover
D3 event handlers are called with two arguments: the event object (e) and
the datum bound to the element that the event was triggered on (d).

- The first thing we do in the handler is select the #info element, because
  all the elements we want to modify are children of that element
- We then update the -a- element inside the -.value- element inside the -.repo-
  element

We're setting both the text content of this element and the 'href' attribute.
- This has the effect of making a link to the repository, with the full name
  of the repository as the link text.

We similarly set the text of the -.value- -.license- element to whatever
-getLicense- returns for this datum and the text of the -.stars- -.value- element
to the number of stars.

For each bar you hover over, the details for that repository should show up in the
new sidebar.

If you click the repository name, your browser should open up a new tab and take you
to the repository's GitHub page.