creating a color scale for licenses
First we collect all the unique license names. -- To do this, we map over the items,
calling our -getLicense- function for each one.

- This gives an array of the license names. In the same line, we pass the resulting
  array to the -Set- constructor.

- A 'set' is a collection of unique items, so the -Set- constructor can take an array
  of items and filter out any duplicates.

In JavaScript sets maintain their order, like arrays.

The -d3.scaleOrdinal- helper creates a scale with discrete inputs and discrete outputs.

- Here the inputs are the unique license names and the outputs are color names.

For the scale's range, we're using -d3.schemeCategory10- which is an array of 10 hex
color strings.

Each license in the set will map to one of these colors, index-wise. -- If there are more
than 10 licenses, the colors will crap around to the beginning again. ( the eleventh and
twelfth licenses will use the same colors as the first and second ones).