Generating the key
Here we're using the #key element as the container for our new join, and we're
joining in a bunch of -p- elements to bind to each license datum.

First we create the -p- element for each new datum, then we append a -div- element
where we'll display the square of color to the -p- element.

Adding the -color- class to the -div- means that it will have the styling from #key
id in style.css

To give it the right color, we use the -style- method, which sets an inline CSS style
on the element. -- We set the color to the appropriate value for the datum using -colorScale-

Finally, we add a -span- element to the -p- element for holding the actual name of the license.

Our visualization now has a full key showing the license that each color maps to,
making the colors much more meaningful.