Adding a left axis label
First, we need to calculate where to draw the label. Its vertical position
should be in the middle of the chart. We calculate the height of the chart
based on the total height and the two margins, then calculate the mid-point
based on adding the top margin and half the height of the chart.

Next, we append the -text- element with the word 'Stars' to the -svg- element.

Setting the -text-anchor- attribute to the 'middle' causes the text to be
centered around its calculated position.

We also specify two transformations: a translate followed by a rotate.
- The 'translate' moves the center of the label to the correct position, and the
-rotate- turns it 90 degress counter-clockwise ( or 270 degrees clockwise ).