Keeping track of the hidden licenses
First, we create a new empty 'Set' called -hiddenLicenses-, just before the
-update- function.

We're using 'Set' here to make it easier to add or remove licenses. -- With
an array it's trickier to remove a specific element.

Then after the code that renders the key, we create a 'change' event handler
for the checkboxes.

Whenever a checkbox changes from checked to unchecked or vice versa, this handler
will run.

In the handler, -e- is the change event and -d- is the bound datum.

We use -e.target.checked- to determine whether, after the change, the checkbox is
checked or not. -- If it is, then we know the datum should be removed from the
-hiddenLicenses- 'set', using the 'delete' method on the set.

Conversely, if the checkbox is now unchecked we add that datum to hiddenLicenses,
using the -add- method on the set.

Finally, with the -hiddenLicenses- set modified, we log the set to the console
and call the -update- function again, with the same items it was called with
originally.

When you reload the page, you won't see any new behavior because we're not actually
updating the graph yet, but if you open the console you'll see how the -hiddenLicenses-
set changes as you check and uncheck the various checkboxes.

Now we need to determine which repositories to show when there are hidden licenses.
To do that, we'll create a new array called -filtered- at the top of the -update-
method.