Determining the repositories with hidden licenses
To filter the list of items, for each item we check to see if its license name is in the
-hiddenLicenses- set, using the set's '.has' method.

If the name isn't included in the set, then it will be in the -filtered- list. Otherwise,
it's filtered out.

Finally, we need to switch to using the -filtered- array rathar than the -items- array
for rendering