Adding animations
Animating the left axis is straightforward: we just call -transistion- on
the axis's container before the axis is drawn on it, and the left axis will
transition anytime the scale changes.

To animate the bars, we follow the standard practice of switching to the
advanced join technique and adding transitions for entering, updating, and
exiting elements.

Entering elements start out with all their attributes set and an opacity of 0
(meaning that they are 100 percent transparent)

We then call -transition- and animate up to 100 percent opaque, which has the
effect of fading in entering elements.

Updating elements remain the same color and opacity, but their position and
dimensions can change, so we animate all of these. -- This has the effect of
stretching and sliding these updated elements to their new size and position.

Exiting elements do the opposite of entering elements and fade out, which we
achieve by transitioning their opacity back to 0

Remember that we also have to call -remove- on any exiting elements after the
transition is complete.

Reload the page the try hiding and unhiding various licenses.

You should see the repositories with that license fade out and the other repositories
expand to fill the licenses down.